Add invoice preview hook
