+++ frontend/glovelly-web/src/App.tsx
@@ -19,8 +19,6 @@ import type { AppNavigationItem, HeaderMetric } from './AppSections'
 import {
   buildApiUrl,
   buildReturnUrl,
-  createBlobObjectUrl,
-  downloadResponseBlob,
   fetchWithSession,
   getResponseErrorMessage,
   isSessionExpiredError,
@@ -36,6 +34,7 @@ import { useAdminWorkspace } from './hooks/useAdminWorkspace'
 import { useClientsWorkspace } from './hooks/useClientsWorkspace'
 import { useGigsWorkspace } from './hooks/useGigsWorkspace'
 import { useGigImportsWorkspace } from './hooks/useGigImportsWorkspace'
+import { useInvoicePreview } from './hooks/useInvoicePreview'
 import { useInvoicesWorkspace } from './hooks/useInvoicesWorkspace'
 import { useProfileMenu } from './hooks/useProfileMenu'
 import { useQuickReceipt } from './hooks/useQuickReceipt'
@@ -86,10 +85,6 @@ function App({ appMetadata }: AppProps) {
   const { setThemePreference, themePreference } = useThemePreference()
   const [monthlyInvoiceMonth, setMonthlyInvoiceMonth] = useState(getCurrentMonthValue)
   const [monthlyInvoiceStatus, setMonthlyInvoiceStatus] = useState('')
-  const [invoicePreviewInvoice, setInvoicePreviewInvoice] = useState<Invoice | null>(null)
-  const [invoicePreviewPdfUrl, setInvoicePreviewPdfUrl] = useState<string | null>(null)
-  const [invoicePreviewStatus, setInvoicePreviewStatus] = useState('')
-  const [isInvoicePreviewLoading, setIsInvoicePreviewLoading] = useState(false)
   const [isGigImportsOpen, setIsGigImportsOpen] = useState(false)
 
   const isAdmin = authUser?.role === 'Admin'
@@ -106,24 +101,6 @@ function App({ appMetadata }: AppProps) {
     [clearSession]
   )
 
-  const clearInvoicePreviewPdfUrl = useCallback(() => {
-    setInvoicePreviewPdfUrl((current) => {
-      if (current) {
-        window.URL.revokeObjectURL(current)
-      }
-
-      return null
-    })
-  }, [])
-
-  useEffect(() => clearInvoicePreviewPdfUrl, [clearInvoicePreviewPdfUrl])
-
-  const closeInvoicePreview = useCallback(() => {
-    clearInvoicePreviewPdfUrl()
-    setInvoicePreviewInvoice(null)
-    setInvoicePreviewStatus('')
-    setIsInvoicePreviewLoading(false)
-  }, [clearInvoicePreviewPdfUrl])
   const {
     activeUsersCount,
     adminForm,
@@ -395,6 +372,17 @@ function App({ appMetadata }: AppProps) {
   } = useSellerProfile({
     onCloseProfileMenu: closeProfileMenu,
     onSessionExpired: expireSession,
+  })
+  const {
+    closeInvoicePreview,
+    downloadInvoicePreviewPdf,
+    invoicePreviewInvoice,
+    invoicePreviewPdfUrl,
+    invoicePreviewStatus,
+    isInvoicePreviewLoading,
+    openInvoicePreview,
+  } = useInvoicePreview({
+    onDownloaded: setInvoiceStatus,
   })
 
   useEffect(() => {
@@ -826,67 +814,6 @@ function App({ appMetadata }: AppProps) {
 
   const closeGigImports = () => {
     setIsGigImportsOpen(false)
-  }
-
-  const openInvoicePreview = async (invoice: Invoice) => {
-    setInvoicePreviewInvoice(invoice)
-    setInvoicePreviewStatus(`Preparing ${invoice.invoiceNumber} preview...`)
-    setIsInvoicePreviewLoading(true)
-    clearInvoicePreviewPdfUrl()
-
-    try {
-      const response = await fetchWithSession(buildApiUrl(`/invoices/${invoice.id}/pdf`))
-
-      if (!response.ok) {
-        throw new Error('Unable to prepare the invoice PDF preview.')
-      }
-
-      const previewUrl = await createBlobObjectUrl(response)
-      setInvoicePreviewPdfUrl((current) => {
-        if (current) {
-          window.URL.revokeObjectURL(current)
-        }
-
-        return previewUrl
-      })
-      setInvoicePreviewStatus(`Invoice ${invoice.invoiceNumber} is ready to review.`)
-    } catch (error) {
-      setInvoicePreviewStatus(
-        error instanceof Error ? error.message : 'Unable to prepare the invoice PDF preview.'
-      )
-    } finally {
-      setIsInvoicePreviewLoading(false)
-    }
-  }
-
-  const downloadInvoicePreviewPdf = async () => {
-    if (!invoicePreviewInvoice) {
-      return
-    }
-
-    const fallbackFilename = `${invoicePreviewInvoice.invoiceNumber}.pdf`
-    setIsInvoicePreviewLoading(true)
-    setInvoicePreviewStatus(`Preparing ${fallbackFilename}...`)
-
-    try {
-      const response = await fetchWithSession(
-        buildApiUrl(`/invoices/${invoicePreviewInvoice.id}/pdf`)
-      )
-
-      if (!response.ok) {
-        throw new Error('Unable to download the invoice PDF.')
-      }
-
-      const filename = await downloadResponseBlob(response, fallbackFilename)
-      setInvoicePreviewStatus(`Downloaded ${filename}.`)
-      setInvoiceStatus(`Downloaded ${filename}.`)
-    } catch (error) {
-      setInvoicePreviewStatus(
-        error instanceof Error ? error.message : 'Unable to download the invoice PDF.'
-      )
-    } finally {
-      setIsInvoicePreviewLoading(false)
-    }
   }
 
   const openPreviewedInvoice = () => {
